renaming file

+++ search_algorithms/0-linear.c
@@ -1,12 +1,14 @@
-#include "search_algm.h"
+#include "search_algos.h"
 /**
+ * linear_search - Function that searches for a value in an array of intergers
+ * usign the Linear search algotithm
  * @array: is a pointer to the first element of the array
  * @size: is the number if elements in array
  * @ value: is the value to search for
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 	for (i = 0; i < size; i++)
 		if (array[i] == value)
 			return i;
